No need for a separate block
