--- assets/blocks/view-certificate/editor.scss
@@ -1,6 +1,0 @@
-.wp-block-sensei-certificates-button-view-certificate .wp-block-button {
-	margin-top: 0;
-	margin-bottom: 0;
-	padding-left: 0;
-	padding-right: 0;
-}
